:recycle: refactor: move responsive to server utils folder
