+++ src/app/(main)/(mobile)/me/settings/page.tsx
@@ -2,7 +2,7 @@ import { redirect } from 'next/navigation';
 
 import { metadataModule } from '@/server/metadata';
 import { translation } from '@/server/translation';
-import { isMobileDevice } from '@/utils/responsive';
+import { isMobileDevice } from '@/utils/server/responsive';
 
 import Category from './features/Category';
