+++ src/app/(main)/chat/(workspace)/page.tsx
@@ -3,7 +3,7 @@ import { BRANDING_NAME } from '@/const/branding';
 import { ldModule } from '@/server/ld';
 import { metadataModule } from '@/server/metadata';
 import { translation } from '@/server/translation';
-import { isMobileDevice } from '@/utils/responsive';
+import { isMobileDevice } from '@/utils/server/responsive';
 
 import PageTitle from '../features/PageTitle';
 import TelemetryNotification from './features/TelemetryNotification';
