+++ src/app/(main)/discover/(list)/models/[slug]/page.tsx
@@ -8,7 +8,7 @@ import { metadataModule } from '@/server/metadata';
 import { DiscoverService } from '@/server/services/discover';
 import { translation } from '@/server/translation';
 import { AssistantCategory } from '@/types/discover';
-import { isMobileDevice } from '@/utils/responsive';
+import { isMobileDevice } from '@/utils/server/responsive';
 
 import List from '../features/List';
