+++ src/app/(main)/discover/search/page.tsx
@@ -7,7 +7,7 @@ import { Locales } from '@/locales/resources';
 import { ldModule } from '@/server/ld';
 import { metadataModule } from '@/server/metadata';
 import { translation } from '@/server/translation';
-import { isMobileDevice } from '@/utils/responsive';
+import { isMobileDevice } from '@/utils/server/responsive';
 
 import { ListLoadingWithoutBanner } from '../components/ListLoading';
