+++ src/app/(main)/profile/layout.tsx
@@ -2,7 +2,7 @@ import { notFound } from 'next/navigation';
 import { PropsWithChildren } from 'react';
 
 import { enableClerk } from '@/const/auth';
-import { isMobileDevice } from '@/utils/responsive';
+import { isMobileDevice } from '@/utils/server/responsive';
 
 import MobileLayout from './_layout/Mobile';
